backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
transbank-sdk
python-multipart

